🐛 Fix deadlock issue caused by blocking queue
The problem is that the main thread (i.e. TCP server) would wait on
joining all the working thread when being asked to stop, whereas the
worker threads would typically deal with the close event of TCP
connections and emit functors to the main thread's blocking queue, whose
size might reach the capacity and cause deadlock since the main thread
is blocked.

And the solution is to make the joining of the worker threads pollable
in the case of using blocking queue (in the case of lock-free queue
there is no limit on the capacity so it doesn't matter whether pollable
or not). To achieve this the thread must be able to sleep (i.e.
`this_thread::sleep_for()`), the TCP server must be able to invoke the
external event loop's functors (i.e. `EventLoop::invoke_all_functors()`)
and be able to poll for the joining status of the worker threads (i.e.
`Thread::is_joinable()`).

+++ include/tcp_server.h
@@ -91,7 +91,7 @@ private:
     // for tcp connect socketfd
     void _close_callback(
         TcpConnectSocketfd *tcp_connect_socketfd_ptr,
-        uint64_t functor_blocking_queue_index
+        size_t functor_blocking_queue_index
     );
 
     void _transfer_tcp_connections_to_background_thread();
